+++ src/main/java/com/example/demo/Service/PpTaskDispatchService.java
@@ -2,13 +2,15 @@ package com.example.demo.Service;
 
 import com.example.demo.Data.PpTaskClaim;
 
+import java.math.BigInteger;
+
 public interface PpTaskDispatchService {
 
     PpTaskClaim claimOne(String deviceId,String deviceNickName , String cardNo);
 
-    void finishSuccess(Long claimId, String deviceId);
+    void finishSuccess(BigInteger claimId, String deviceId);
 
-    void finishFail(Long claimId, String deviceId,String msg);
+    void finishFail(BigInteger claimId, String deviceId, String msg);
 
     void expireOverdueClaims();
